Refactor author initialization
Instead of each author having the whole list of naughty words (451
words) initialized on creation, we now only insert the swear words when
they are encountered in the repository. This was a surprisingly big
win in run time, on a repo with ~500 commits the total run time went
from ~0.4s to ~0.05s. Neat.