More suggestions from #1
Add authors to the repo struct, I was silly and thought it was
reasonable to put them in a HashSet but then I couldn't retrieve them
because I had nothing to match them against. Using a HashMap where the
key is what I should've seen... but I didn't.

Add function that lazily adds authors to the repo, a better structure
of the main loop and a much better way of adding and retrieving authors.